Backend-flask/requirements.txt
flask
flask-cors
psycopg2-binary
celery
redis
